fix: correct minor issues found during testing
Clean client IP address

Server password policy strengthened

Username and Realname min lenght applied - Their first char cannot be a digit

Handle welcome message

Adds empty LIST command

Command LIST complete

Refactor : Server split

Broadcast message for Op promotion

Member to Op promotion sends the appropriate message
