+++ incs/Server.hpp
@@ -6,7 +6,7 @@
 /*   By: mhotting <mhotting@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/31 17:14:57 by mhotting          #+#    #+#             */
-/*   Updated: 2025/10/07 23:11:36 by mhotting         ###   ########.fr       */
+/*   Updated: 2025/10/08 18:38:41 by mhotting         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,52 +23,56 @@
 #include <vector>
 
 class Command;
+class Channel;
 
 class Server {
 public:
+    // Server general
     Server(int port, const std::string &name, const std::string &password, const std::string &version);
     ~Server(void);
+    const std::string &getName(void) const;
+    void closeFds(void);
+    void init(void);
+    static void signalHandler(int);
+    bool isValidPassword(const std::string &password);
 
+    // Server network
     int getSocketFd(void) const;
     int getPort(void) const;
-    const std::string &getName(void) const;
-    Client *getClientByFd(int fd);
-    Client *getClientByNickname(const std::string &nick);
-    const std::vector<Client *> &getAllClients(void) const;
-    std::vector<std::string> getAllChannelNames(void) const;
-
-    void init(void);
     void createSocket(void);
-    void acceptNewClient(void);
     void receiveData(int fd);
     void sendData(Client *client);
 
-    void closeFds(void);
+    // Server client
+    Client *getClientByFd(int fd);
+    Client *getClientByNickname(const std::string &nick);
+    const std::vector<Client *> &getAllClients(void) const;
+    void acceptNewClient(void);
     void clearClient(int fd, const std::string &reason = "");
     void markClientForWrite(int clientFd);
-
-    static void signalHandler(int);
-
-    bool isValidPassword(const std::string &password);
     bool isNicknameInUse(const std::string &nick);
 
+    // Server channel
+    std::vector<std::string> getAllChannelNames(void) const;
+    const std::vector<Channel> &getChannels(void) const;
     Channel &addChannel(const std::string &channelName);
     Channel *getChannelByName(const std::string channelName);
     bool isChannelCreated(const std::string channelName);
     void removeChannel(const std::string &channelName);
-
     void partClientFromAllChannels(Client *client, const std::string &reason, const std::string &commandName = IRC::CMD_PART);
     void partClientFromChannel(Client *client, Channel &channel, const std::string &reason, const std::string &commandName = IRC::CMD_PART);
     void quitClientFromAllChannels(Client *client, const std::string &reason);
+    std::set<Client *> getChannelPeers(Client *client);
 
+    // Server messages
     void registerClient(Client *client);
-
     void sendNumericReplyToClient(Client *client, int code, const std::string &message);
     void sendNumericReplyToClient(Client *client, int code, const std::string &message, const std::string &param);
+    void sendMessageToClient(Client *dest_client, const std::string &commandName, const std::string &params);
     void sendMessageToClient(const Client *source_client, Client *dest_client, const std::string &commandName, const std::string &params);
     void sendMessageToChannelUsers(Channel &channel, const Client *source, const std::string &commandName, const std::string &params);
+    void sendMessageToChannelUsers(Channel &channel, const std::string &commandName, const std::string &params);
     void sendNamesToClient(Client *client, Channel &channel);
-    std::set<Client *> getChannelPeers(Client *client);
 
 private:
     int _socketFd;
@@ -76,6 +80,7 @@ private:
     const std::string _name;
     const std::string _password;
     const std::string _version;
+    std::string _serverIp;
     std::string _creationDate;
     std::vector<Client *> _clients;
     std::vector<Channel> _channels;
@@ -83,11 +88,14 @@ private:
     static bool _signalReceived;
     std::queue<Command *> _commandQueue;
 
+    // Server general
     Server(void);
-    void _addClientToPollList(int fd);
-
     void _extractCommands(void);
     void _processCommands(void);
+    void _printWelcomeMessage(void) const;
+
+    // Server client
+    void _addClientToPollList(int fd);
 };
 
 #endif
