+++ incs/commands/ListCommand.hpp
@@ -1,0 +1,30 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ListCommand.hpp                                    :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: mhotting <mhotting@student.42lyon.fr>      +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/10/08 15:28:09 by mhotting          #+#    #+#             */
+/*   Updated: 2025/10/08 15:29:26 by mhotting         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef LISTCOMMAND_HPP
+
+#include "Command.hpp"
+
+class ListCommand : public Command {
+public:
+    static const std::string NAME;
+    ListCommand(Client *sender, const std::vector<std::string> &params);
+
+    const std::string &getName(void) const;
+
+    void execute(Server &server);
+
+private:
+    ListCommand(Client *client);
+};
+
+#endif
