+++ incs/helpers.hpp
@@ -6,7 +6,7 @@
 /*   By: mhotting <mhotting@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/31 18:33:47 by mhotting          #+#    #+#             */
-/*   Updated: 2025/10/02 14:44:04 by mhotting         ###   ########.fr       */
+/*   Updated: 2025/10/08 18:32:17 by mhotting         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,6 +24,7 @@ void validatePassword(const std::string &password);
 std::string formatNumericReply(const std::string &serverName, int code, const std::string &target, const std::string &message);
 std::string formatNumericReply(const std::string &serverName, int code, const std::string &target, const std::string &param, const std::string &message);
 std::string formatMessage(const Client *client, const std::string &command_name, const std::string &params);
+std::string formatMessage(const std::string &formattedName, const std::string &command_name, const std::string &params);
 
 bool isLetter(char c);
 bool isDigit(char c);
