+++ srcs/main.cpp
@@ -6,7 +6,7 @@
 /*   By: mhotting <mhotting@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/03 15:08:42 by mhotting          #+#    #+#             */
-/*   Updated: 2025/09/30 17:30:26 by mhotting         ###   ########.fr       */
+/*   Updated: 2025/10/08 12:20:22 by mhotting         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,7 +39,6 @@ int main(int argc, char **argv) {
     }
 
     // Server creation
-    std::cout << "ft_irc - port: " << port << " - password: " << password << std::endl;
     Server server(port, SERVER_NAME, password, SERVER_VERSION);
     try {
         signal(SIGINT, Server::signalHandler);
